backend/requirements.txt
aniso8601==6.0.0
Click==7.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.7
Flask-SQLAlchemy==3.0.3
Flask-Session==0.4.0s
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2-binary==3.1.8
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==0.15.5
psycopg2==2.9.6
python-dotenv==1.0.0